Factor out common code between benchmarks and tests
